rag-scheme-main/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
pandas==2.1.4
numpy==1.24.3
huggingface_hub==0.19.4
faiss-cpu==1.7.4
pydantic==2.5.0
google-generativeai==0.7.2
langchain-google-genai==2.0.0